+++ src/client/app/humidity-sensor/humidity-sensor-view-model.spec.ts
@@ -1,4 +1,4 @@
-import {HumiditySensorViewModel} from './humiditySensor-view-model';
+import {HumiditySensorViewModel} from './humidity-sensor-view-model';
 import {MiscUtils} from '../shared/global/misc-utils';
 export function main() {
   let humiditySensorsViewModel: HumiditySensorViewModel;
