+++ src/client/app/humidity-sensor/humidity-sensor-view-model.ts
